+++ src/components/navigation.tsx
@@ -1,0 +1,8 @@
+"use client";
+export default function Navigation() {
+  return (
+    <nav>
+      <ul></ul>
+    </nav>
+  );
+}
